add checks after calling getHostnames(args)
This prevents a class of bugs where a valid scope or wildcard selector is in `args` but resolves to no matches

```
args = ['a:empty_appliance', 'unmatching_regex*']
hosts = self.getHostnames(args)
for host in hosts:
  pass

self.call('some.command', hosts)

```

+++ common/src/stack/command/stack/commands/remove/host/partition/__init__.py
@@ -57,8 +57,12 @@ class Command(stack.commands.remove.host.command):
 			('partition', None),
 			('device', None),
 			('uuid', None)])
-			
-		for host in self.getHostnames(args):
+
+		hosts = self.getHostnames(args)
+		if not hosts:
+			return
+
+		for host in hosts:
 			conditions = []
 			sql_cmd = """delete from partitions where
 				node=(select id from nodes
